requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
braceexpand==0.1.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
frozenlist==1.3.0
idna==3.3
import-expression==1.1.4
jishkucord==2.5.2
motor==3.0.0
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
prettytable==3.3.0
py-cord==2.0.0rc1
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
wcwidth==0.2.5
yarl==1.7.2
